Fixed readme, added release to releases tab in remote repo
