+++ build.gradle
@@ -34,7 +34,7 @@ sourceSets {
 
 intellij {
     version 'IC-15.0.5'
-    pluginName 'PerfomanceMetricsPlugin'
+    pluginName 'ProductivityMetricsPlugin'
     updateSinceUntilBuild true
     downloadSources true
 }
