+++ src/net/svitkov/ProductivityMetricsPlugin/handlers/BackspaceHandler.java
@@ -15,26 +15,30 @@
  * limitations under the License.
  */
 
-package net.svitkov.PerfomanceMetricsPlugin.handlers;
+package net.svitkov.ProductivityMetricsPlugin.handlers;
 
-import com.intellij.codeInsight.editorActions.TypedHandlerDelegate;
+import com.intellij.codeInsight.editorActions.BackspaceHandlerDelegate;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.fileEditor.FileDocumentManager;
-import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiFile;
-import org.jetbrains.annotations.NotNull;
-import net.svitkov.PerfomanceMetricsPlugin.stuff.Counter;
+import net.svitkov.ProductivityMetricsPlugin.stuff.Counter;
 
-public class CountingHandler extends TypedHandlerDelegate {
+public class BackspaceHandler extends BackspaceHandlerDelegate{
 
-    @Override
-    public Result charTyped(char c, final Project project, final @NotNull Editor editor, @NotNull final PsiFile file) {
-        Counter counter = Counter.getInstance();
-	    Document currentDocument = editor.getDocument();
-	    String openedFileName = FileDocumentManager.getInstance().getFile(currentDocument).getName();
-	    ApplicationManager.getApplication().runReadAction(() -> counter.increment(openedFileName));
-        return Result.CONTINUE;
-    }
+	@Override
+	public void beforeCharDeleted(char c, PsiFile file, Editor editor) {
+
+	}
+
+	@Override
+	public boolean charDeleted(char c, PsiFile file, Editor editor) {
+		// Probably could be simplified
+		Counter counter = Counter.getInstance();
+		Document currentDocument = editor.getDocument();
+		String openedFileName = FileDocumentManager.getInstance().getFile(currentDocument).getName();
+		ApplicationManager.getApplication().runReadAction(() -> counter.decrement(openedFileName));
+		return false;
+	}
 }
